Close browser at the end of final test run
When I was using this setup Chromium was not closing after the final test run and was being interrupted when the parent Jest process quit. This fixed it for me. I'd be interested to know if anyone else had these problems or whether there is another canonical way of solving this problem!